requirements.txt
adal
adlfs
aiobotocore
aiofiles
aiohttp
aioitertools
aiosignal
alembic
apache-airflow-providers-amazon==8.27.0
apache-airflow-providers-celery==3.5.0
apache-airflow-providers-cncf-kubernetes==7.11.0
apache-airflow-providers-common-compat==1.1.0
apache-airflow-providers-common-io==1.1.0
apache-airflow-providers-common-sql==1.9.0
apache-airflow-providers-docker==3.8.2
apache-airflow-providers-elasticsearch==5.3.0
apache-airflow-providers-ftp==3.7.0
apache-airflow-providers-google==10.12.0
apache-airflow-providers-grpc==3.4.0
apache-airflow-providers-hashicorp==3.6.0
apache-airflow-providers-http==4.8.0
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-microsoft-azure==8.4.0
apache-airflow-providers-mysql==5.5.0
apache-airflow-providers-odbc==4.2.0
apache-airflow-providers-openlineage==1.3.0
apache-airflow-providers-postgres==5.9.0
apache-airflow-providers-redis==3.5.0
apache-airflow-providers-sendgrid==3.4.0
apache-airflow-providers-sftp==4.8.0
apache-airflow-providers-slack==8.5.0
apache-airflow-providers-snowflake==5.2.0
apache-airflow-providers-sqlite==3.6.0
apache-airflow-providers-ssh==3.9.0
amqp
annotated-types
anyio
apispec
argcomplete
asgiref
asn1crypto
async-timeout
attrs
Authlib
Babel
backoff
bcrypt
beautifulsoup4
billiard
blinker
boto3
botocore
cachelib
cachetools
cattrs
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
clickclick
colorama
colorlog
ConfigUpdater
connexion
cron-descriptor
croniter
cryptography
db-dtypes
decorator
Deprecated
dill
distlib
dnspython
docker
docutils
elastic-transport
elasticsearch
email-validator
et-xmlfile
eventlet
exceptiongroup
filelock
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
flower
frozenlist
fsspec
gcloud-aio-auth
gcloud-aio-bigquery
gcloud-aio-storage
gcsfs
gevent
google-ads
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-crc32c
google-re2
google-resumable-media
googleapis-common-protos
graphviz
greenlet
grpc-google-iam-v1
grpcio
grpcio-gcp
grpcio-status
gunicorn
h11
httpcore
httplib2
httpx
humanize
hvac
idna
ijson
importlib-metadata
importlib-resources
inflection
isodate
itsdangerous
Jinja2
jmespath
json-merge-patch
jsonpath-ng
jsonschema
jsonschema-specifications
kombu
kubernetes
kubernetes-asyncio
lazy-object-proxy
ldap3
limits
linkify-it-py
lockfile
looker-sdk
lxml
Mako
Markdown
markdown-it-py
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
msal
msal-extensions
msrest
msrestazure
multidict
mysql-connector-python
numpy
oauthlib
Office365-REST-Python-Client
openlineage-integration-common
openlineage-python
openlineage_sql
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
packaging
pandas
pandas-gbq
paramiko
pathspec
pendulum
platformdirs
pluggy
ply
portalocker
prison
prometheus-client
prompt-toolkit
proto-plus
protobuf
psutil
psycopg2-binary
pyarrow
pyarrow-hotfix
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic_core
pydata-google-auth
Pygments
PyJWT
PyNaCl
pyodbc
pyOpenSSL
pyparsing
pyspnego
python-daemon
python-dateutil
python-dotenv
python-http-client
python-nvd3
python-slugify
pytz
pytzdata
PyYAML
redis
redshift-connector
referencing
requests
requests-ntlm
requests-oauthlib
requests-toolbelt
rfc3339-validator
rfc3986
rich
rich-argparse
rpds-py
rsa
s3transfer
scramp
sendgrid
setproctitle
shapely
six
slack-sdk
sniffio
snowflake-connector-python
snowflake-sqlalchemy
sortedcontainers
soupsieve
SQLAlchemy
sqlalchemy-bigquery
SQLAlchemy-JSONField
sqlalchemy-redshift
sqlalchemy-spanner
SQLAlchemy-Utils
sqlparse
sshtunnel
starkbank-ecdsa
statsd
tabulate
tenacity
termcolor
text-unidecode
tomlkit
tornado
typing_extensions
tzdata
uc-micro-py
unicodecsv
universal-pathlib
uritemplate
urllib3
vine
virtualenv
watchtower
wcwidth
websocket-client
Werkzeug
wrapt
WTForms
yarl
zipp
zope.event
zope.interface
zipp
zope.event
zope.interface
azure-core
azure-storage-blob
azure-mgmt-resource
azure-mgmt-storage
azure-identity
azure-keyvault-secrets
azure-cosmos
azure-data-tables
pendulum
apache-airflow-providers-amazon